# [CR needed] STFT pre-allocation