#### Reference Issue
Fixes #871 


#### What does this implement/fix? Explain your changes.

This PR adds parameter `out=` to `stft` and `istft` methods, allowing re-use of previously allocated output buffers.

Additionally, it implements some more careful handling of edge padding (when using centered mode) to reduce unnecessary copying.  The result should be a substantial speedup when processing long signals.

#### Any other comments?

This is currently WIP.  To-dos:

- [x] complete tests for stft pre-allocation
- [x] add istft and tests
- [x] update other parts of the library to make use of pre-allocation where possible, eg griffinlim
- [x] add an example to the gallery illustrating the utility, or maybe just extend the pcen example

That said, I'd like to get some early feedback on the implementation, mainly from the perspective of readability.