@bmcfee 2022-06-24
And we've hit a pretty serious snag that I hadn't considered.  The optimization I've implemented to only pad at the edges breaks padding semantics for `wrap` mode.  This isn't a huge deal in the big picture, but in the short term it does break some of our unit tests (eg on reassigned frequencies).

Now, I'm not entirely sure why we're using wrap padding in these tests - probably this is to get around edge transients for the stationary test signals, but it's a pretty crude hack.

However, given that this does change semantics in a pretty subtle way, I don't think we can include this optimization in a minor point release.  We do have the option of rolling back to just have pre-allocated output arrays without the copy-pad optimization, but my preference would be to keep this intact and bump it to a later release (eg 0.10).