@bmcfee 2022-07-01
I've added exceptions for now-unsupported padding modes in stft.

I wound up disabling the `center=True` test for reassigned spectrograms for now.  We used `pad_mode='wrap'` not out of actual necessity, but just to minimize the influence of edge transients on our test case.  Ultimately this was a bit of a hack, and I don't think it diminishes much to simplify this test and only check `center=False` (no padding).

The remaining to-do's in the issue are still TBD though.