@bmcfee 2022-07-02
Prototype of griffin-lim using pre-allocation for both forward and inverse STFT's gets us a minor boost.  Example is 90 seconds of random noise at float32.

Here, `librosa.griffinlim` is not using pre-allocation, but is using the optimized padding from this PR.  `griffinlim` is using pre-allocation.  Results are numerically identical:

```python
>>> sr = 22050
>>> y = np.random.randn(22050 * 90).astype(np.float32)
>>> %timeit librosa.griffinlim(S, random_state=0, n_iter=64)
16.4 s ± 1.64 s per loop (mean ± std. dev. of 7 runs, 1 loop each)
>>> %memit librosa.griffinlim(S, random_state=0, n_iter=64)
peak memory: 518.35 MiB, increment: 151.34 MiB
>>> %timeit griffinlim(S, random_state=0, n_iter=64)
15.1 s ± 370 ms per loop (mean ± std. dev. of 7 runs, 1 loop each)
>>> %memit griffinlim(S, random_state=0, n_iter=64)
peak memory: 518.42 MiB, increment: 151.40 MiB
```
Peak memory hasn't changed much, which I think just means the gc was doing a good job previously of cleaning up after itself.

At this point, I think we're still paying primarily in temp storage from the intermediate reconstruction here:
https://github.com/librosa/librosa/blob/9f46e987488a98ad383aef2875e95ae970c057e5/librosa/core/spectrum.py#L2434
we could probably absorb this directly into the `angles` array with an in-place multiply and cut down on storage a bit more.