@bmcfee 2022-07-02
I had to add a couple of additional tweaks to GL here, mainly to preserve numerical precision in angles dtype.

Previously we forced complex64 even if the input was higher precision.  When moving from a temporary multiply (`istft(S * angles, ...)`) to an in-place multiply (`angles *= S`), this was causing disagreements on double-precision inputs (single-precision retained numerical equivalence).  This PR now has GL adapting the complex dtype to the precision of the input (`S`).

I'm still not seeing a ton of reliable speed improvements here, though we should automatically inherit the previously reported speedups in stft.