@bmcfee 2022-07-02
> I'm still not seeing a ton of reliable speed improvements here, though we should automatically inherit the previously reported speedups in stft.

Quick follow-up: I am seeing reliable memory improvements though, so I think it's an optimization worth taking.

0.9.2, 10-second noise at 22050:

```python
>>> %memit librosa.griffinlim(S, random_state=0, n_iter=64)  # 0.9.2 branch
peak memory: 267.21 MiB, increment: 14.00 MiB
>>> %memit griffinlim(S, random_state=0, n_iter=64)  # PR 1514
peak memory: 198.05 MiB, increment: 14.25 MiB
```