@bmcfee 2022-07-02
Ok, I think this one is feature-complete.  It does have some fairly substantial changes though.  While everything here "should be" numerically equivalent to the 0.9.2 branch (notwithstanding the previous comment about numerical precision in griffin-lim phase estimates), I think this PR does warrant some careful scrutiny by someone apart from myself.