@bmcfee 2022-07-08
Some additional benchmarking data, relevant to #599.  I've been playing around with asv to get a more long-term sense of efficiency gains and regressions.  Here are some benchmark results for time and peak memory (on my laptop) for the current main branch vs this PR when running STFT on data of different durations (in seconds), different hop lengths, and centering on/off:

main branch (before preallocation optimizations):
```
[  0.00%] · For librosa commit 25538adb <abs2>:
[  0.00%] ·· Benchmarking conda-py3.9
[ 50.00%] ··· Running (benchmarks.STFTSuite.time_stft--).
[ 75.00%] ··· benchmarks.STFTSuite.peakmem_stft                                                                                                                                             ok
[ 75.00%] ··· ============ =========== =========== ============ ============ ============ =============
              --                                        center / duration                              
              ------------ ----------------------------------------------------------------------------
               hop_length   True / 10   True / 60   True / 240   False / 10   False / 60   False / 240 
              ============ =========== =========== ============ ============ ============ =============
                  256          154M        193M        336M         153M         188M          315M    
                  512          150M        172M        251M         149M         167M          230M    
                  1024         148M        162M        209M         148M         156M          188M    
              ============ =========== =========== ============ ============ ============ =============

[100.00%] ··· benchmarks.STFTSuite.time_stft                                                                                                                                                ok
[100.00%] ··· ============ ============= ============ ============ ============= ============ =============
              --                                          center / duration                                
              ------------ --------------------------------------------------------------------------------
               hop_length    True / 10    True / 60    True / 240    False / 10   False / 60   False / 240 
              ============ ============= ============ ============ ============= ============ =============
                  256       10.6±0.07ms   71.7±20ms     491±10ms     10.8±0.3ms   67.2±20ms      453±10ms  
                  512        5.81±0.2ms   31.3±0.4ms    253±20ms     5.47±0.1ms   31.2±0.8ms     205±7ms   
                  1024       3.65±0.4ms   18.5±0.7ms    132±6ms     3.09±0.05ms   17.1±0.5ms    77.5±20ms  
              ============ ============= ============ ============ ============= ============ =============
```
PR 1514:
```
[  0.00%] · For librosa commit c95314eb <stft-preallocate>:
[  0.00%] ·· Benchmarking conda-py3.9
[ 50.00%] ··· Running (benchmarks.STFTSuite.time_stft--).
[ 75.00%] ··· benchmarks.STFTSuite.peakmem_stft                                                                                                                                             ok
[ 75.00%] ··· ============ =========== =========== ============ ============ ============ =============
              --                                        center / duration                              
              ------------ ----------------------------------------------------------------------------
               hop_length   True / 10   True / 60   True / 240   False / 10   False / 60   False / 240 
              ============ =========== =========== ============ ============ ============ =============
                  256          153M        188M        315M         153M         188M          315M    
                  512          149M        167M        230M         149M         167M          230M    
                  1024         148M        156M        188M         148M         156M          188M    
              ============ =========== =========== ============ ============ ============ =============

[100.00%] ··· benchmarks.STFTSuite.time_stft                                                                                                                                                ok
[100.00%] ··· ============ ============= ============ ============ ============= ============ =============
              --                                          center / duration                                
              ------------ --------------------------------------------------------------------------------
               hop_length    True / 10    True / 60    True / 240    False / 10   False / 60   False / 240 
              ============ ============= ============ ============ ============= ============ =============
                  256        11.6±0.6ms   74.1±0.5ms    291±1ms      11.3±0.3ms   74.0±0.3ms     292±2ms   
                  512       5.94±0.02ms   33.1±0.8ms   149±0.9ms    5.72±0.09ms   31.8±0.8ms     149±1ms   
                  1024      3.32±0.04ms   17.6±0.6ms   77.8±0.8ms    3.24±0.1ms   17.7±0.7ms    77.4±0.9ms 
              ============ ============= ============ ============ ============= ============ =============
```

At some point, I'll have this up to run continuously on a server and publish results on a website.  But for now, it at least gives a rough sense of time and memory improvements for this PR.